feat(the-platform): wire Stepper into page.tsx, delete Timeline files
