+++ sites/arolariu.ro/src/app/about/the-platform/_components/Stepper.module.scss
@@ -1,0 +1,214 @@
+@use '../../../../styles/abstracts' as abstracts;
+
+.section {
+  position: relative;
+  padding: abstracts.space(24) 0;
+}
+
+.container {
+  max-width: 48rem;
+  margin-left: auto;
+  margin-right: auto;
+  padding: 0 abstracts.space(4);
+}
+
+.header {
+  margin-bottom: abstracts.space(16);
+  text-align: center;
+}
+
+.badge {
+  margin-bottom: abstracts.space(4);
+}
+
+.title {
+  margin-bottom: abstracts.space(6);
+  font-size: abstracts.font-size('4xl');
+  font-weight: abstracts.font-weight('bold');
+
+  @include abstracts.respond-to('md') {
+    font-size: abstracts.font-size('5xl');
+  }
+}
+
+.titleHighlight {
+  color: abstracts.color('primary');
+}
+
+.description {
+  font-size: abstracts.font-size('lg');
+  color: abstracts.color('muted-foreground');
+}
+
+.stepper {
+  position: relative;
+}
+
+.step {
+  display: flex;
+  gap: abstracts.space(4);
+}
+
+.stepLeft {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  flex-shrink: 0;
+  width: 2.5rem;
+}
+
+.stepIcon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 2.5rem;
+  height: 2.5rem;
+  border-radius: abstracts.radius('full');
+  background-color: abstracts.color('primary');
+  flex-shrink: 0;
+}
+
+.stepIconInner {
+  width: 1.25rem;
+  height: 1.25rem;
+  color: white;
+}
+
+.connector {
+  width: 2px;
+  flex: 1;
+  min-height: abstracts.space(8);
+  background-color: abstracts.color('border');
+  margin-top: abstracts.space(2);
+  margin-bottom: abstracts.space(2);
+}
+
+.stepContent {
+  flex: 1;
+  min-width: 0;
+  padding-bottom: abstracts.space(8);
+}
+
+.dateBadge {
+  margin-bottom: abstracts.space(2);
+  display: inline-flex;
+  align-items: center;
+  gap: abstracts.space(1.5);
+}
+
+.dateIcon {
+  width: 0.875rem;
+  height: 0.875rem;
+}
+
+.stepButton {
+  background: transparent;
+  border: none;
+  padding: 0;
+  text-align: left;
+  cursor: pointer;
+  width: 100%;
+}
+
+.stepTitle {
+  font-size: abstracts.font-size('xl');
+  font-weight: abstracts.font-weight('semibold');
+  color: abstracts.color('foreground');
+  margin-bottom: abstracts.space(2);
+  @include abstracts.transition(color);
+
+  .stepButton:hover & {
+    color: abstracts.color('primary');
+  }
+}
+
+.stepDescription {
+  color: abstracts.color('muted-foreground');
+  margin-bottom: abstracts.space(3);
+}
+
+.tags {
+  display: flex;
+  flex-wrap: wrap;
+  gap: abstracts.space(2);
+  margin-bottom: abstracts.space(3);
+}
+
+.tag {
+  font-size: abstracts.font-size('xs');
+}
+
+.expandable {
+  overflow: hidden;
+  margin-bottom: abstracts.space(2);
+}
+
+.detailsTitle {
+  font-size: abstracts.font-size('sm');
+  font-weight: abstracts.font-weight('semibold');
+  margin-bottom: abstracts.space(2);
+  color: abstracts.color('foreground');
+}
+
+.detailsList {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  gap: abstracts.space(1.5);
+}
+
+.detailItem {
+  display: flex;
+  align-items: flex-start;
+  gap: abstracts.space(2);
+  font-size: abstracts.font-size('sm');
+  color: abstracts.color('muted-foreground');
+}
+
+.detailIcon {
+  width: 1rem;
+  height: 1rem;
+  color: abstracts.color('primary');
+  flex-shrink: 0;
+  margin-top: 0.125rem;
+}
+
+.expandHint {
+  font-size: abstracts.font-size('sm');
+  color: abstracts.color('muted-foreground');
+  cursor: pointer;
+}
+
+.futureIndicator {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: abstracts.space(3);
+  margin-top: abstracts.space(8);
+  padding-top: abstracts.space(8);
+  border-top: 1px solid abstracts.color('border');
+  text-align: center;
+}
+
+.futureIconWrapper {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 3rem;
+  height: 3rem;
+  border-radius: abstracts.radius('full');
+  background-color: abstracts.color-alpha('primary', 0.1);
+}
+
+.futureIcon {
+  width: 1.5rem;
+  height: 1.5rem;
+  color: abstracts.color('primary');
+}
+
+.futureText {
+  color: abstracts.color('muted-foreground');
+  font-size: abstracts.font-size('sm');
+}
