--- sites/arolariu.ro/src/app/about/the-platform/_components/Timeline.stories.tsx
@@ -1,22 +1,0 @@
-import type {Meta, StoryObj} from "@storybook/react";
-import Timeline from "./Timeline";
-
-/**
- * Timeline section displaying the platform's development history.
- * Features an interactive alternating timeline with eight milestone events,
- * expandable detail sections, technology tags, and a future indicator.
- * Uses the `About.Platform.timeline` i18n namespace.
- */
-const meta = {
-  title: "Pages/About/ThePlatform/Timeline",
-  component: Timeline,
-  parameters: {
-    layout: "fullscreen",
-  },
-} satisfies Meta<typeof Timeline>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-/** Default timeline with eight expandable milestone events. */
-export const Default: Story = {};
