+++ sites/arolariu.ro/src/app/about/the-platform/page.tsx
@@ -18,7 +18,7 @@ import CallToAction from "./_components/CallToAction";
 import Features from "./_components/Features";
 import Hero from "./_components/Hero";
 import TechStack from "./_components/TechStack";
-import Timeline from "./_components/Timeline";
+import Stepper from "./_components/Stepper";
 import styles from "./page.module.scss";
 
 /**
@@ -86,8 +86,8 @@ export default async function AboutThePlatformPage(_props: Readonly<PageProps<"/
         {/* Tech Stack Section - Technologies used (with merged statistics) */}
         <TechStack />
 
-        {/* Timeline Section - Development journey */}
-        <Timeline />
+        {/* Stepper Section - Development journey */}
+        <Stepper />
 
         {/* Call to Action Section - Footer CTA */}
         <CallToAction />
